requirements.txt
pandas
psycopg2-binary
transformers==4.31.0
python-dotenv
torch
tiktoken==0.4.0
protobuf
pytest